+++ extensions/wrapper/wrapper-common-api/src/main/java/de/sovity/edc/ext/wrapper/api/common/model/UiAsset.java
@@ -38,9 +38,6 @@ public class UiAsset {
     @Schema(description = "Asset Id", requiredMode = Schema.RequiredMode.REQUIRED)
     private String id;
 
-    @Schema(description = "Asset Name", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    private String name;
-
     @Schema(description = "Asset Title", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     private String title;
 
@@ -51,10 +48,10 @@ public class UiAsset {
     private String description;
 
     @Schema(description = "Asset Organization Name", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    private String creator;
+    private String creatorOrganizationName;
 
     @Schema(description = "Asset Homepage", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    private String publisher;
+    private String publisherHomepage;
 
     @Schema(description = "License URL", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     private String licenseUrl;
@@ -68,7 +65,7 @@ public class UiAsset {
     @Schema(description = "Asset MediaType", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     private String distribution;
 
-    @Schema(description = "Landing Page URL", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
+    @Schema(description = "Homepage URL associated with the Asset", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     private String landingPageUrl;
 
     @Schema(description = "HTTP Datasource Hints Proxy Method", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
@@ -106,9 +103,6 @@ public class UiAsset {
 
     @Schema(description = "Asset Json Properties", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     private Map<String, String> additionalJsonProperties;
-
-    @Schema(description = "Asset DataAddress Properties", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    private Map<String, String> dataAddressProperties;
 
     @Schema(description = "Contains the entire asset in its original expanded JSON-LD format", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     private String assetJsonLd;
