+++ extensions/wrapper/wrapper-common-api/src/main/java/de/sovity/edc/ext/wrapper/api/common/model/UiAssetCreateRequest.java
@@ -32,9 +32,6 @@ public class UiAssetCreateRequest {
     @Schema(description = "Asset Id", requiredMode = Schema.RequiredMode.REQUIRED)
     private String id;
 
-    @Schema(description = "Asset Name", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    private String name;
-
     @Schema(description = "Asset Title", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     private String title;
 
@@ -45,10 +42,10 @@ public class UiAssetCreateRequest {
     private String description;
 
     @Schema(description = "Asset Organization Name", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    private String creator;
+    private String creatorOrganizationName;
 
     @Schema(description = "Asset Homepage", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    private String publisher;
+    private String publisherHomepage;
 
     @Schema(description = "License URL", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     private String licenseUrl;
