+++ extensions/wrapper/wrapper-common-mappers/src/main/java/de/sovity/edc/ext/wrapper/api/common/mappers/utils/AssetCreatorOrganizationNameJsonLd.java
@@ -1,7 +1,6 @@
 package de.sovity.edc.ext.wrapper.api.common.mappers.utils;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -16,11 +15,7 @@ import lombok.ToString;
 @AllArgsConstructor
 @Builder(toBuilder = true)
 @RequiredArgsConstructor
-@Schema(description = "Asset Creator Details")
-public class UiAssetCreator {
-
-    @JsonProperty("@type")
-    private String type;
+public class AssetCreatorOrganizationNameJsonLd {
 
     @JsonProperty("http://xmlns.com/foaf/0.1/name")
     private String name;
