+++ utils/test-connector-remote/src/main/java/de/sovity/edc/extension/e2e/connector/ConnectorRemote.java
@@ -16,6 +16,7 @@ package de.sovity.edc.extension.e2e.connector;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import de.sovity.edc.extension.e2e.connector.config.ConnectorRemoteConfig;
 import de.sovity.edc.extension.e2e.connector.config.api.auth.NoneAuthProvider;
+import de.sovity.edc.utils.jsonld.vocab.Prop;
 import io.restassured.http.Header;
 import io.restassured.specification.RequestSpecification;
 import jakarta.json.Json;
@@ -38,7 +39,6 @@ import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicReference;
 
-import de.sovity.edc.utils.jsonld.vocab.Prop;
 import static io.restassured.RestAssured.given;
 import static io.restassured.http.ContentType.JSON;
 import static jakarta.json.Json.createObjectBuilder;
@@ -52,10 +52,6 @@ import static org.eclipse.edc.jsonld.spi.PropertyAndTypeNames.DCAT_DATASET_ATTRI
 import static org.eclipse.edc.jsonld.spi.PropertyAndTypeNames.ODRL_POLICY_ATTRIBUTE;
 import static org.eclipse.edc.spi.CoreConstants.EDC_NAMESPACE;
 import static org.eclipse.edc.spi.CoreConstants.EDC_PREFIX;
-import static org.eclipse.edc.spi.types.domain.asset.Asset.PROPERTY_DESCRIPTION;
-import static org.eclipse.edc.spi.types.domain.asset.Asset.PROPERTY_ID;
-import static org.eclipse.edc.spi.types.domain.asset.Asset.PROPERTY_NAME;
-import static org.eclipse.edc.spi.types.domain.asset.Asset.PROPERTY_VERSION;
 
 @SuppressWarnings("java:S5960")
 @RequiredArgsConstructor
@@ -75,7 +71,7 @@ public class ConnectorRemote {
                 .add("asset", createObjectBuilder()
                         .add(ID, assetId)
                         .add("properties", createObjectBuilder()
-                                .add(Prop.DCMI.description, "description")))
+                                .add(Prop.Dcterms.DESCRIPTION, "description")))
                 .add("dataAddress", createObjectBuilder(dataAddressProperties))
                 .build();
 
@@ -211,7 +207,7 @@ public class ConnectorRemote {
                 .add("protocol", "dataspace-protocol-http")
                 .add("offer", createObjectBuilder()
                         .add("offerId", offerId)
-                        .add("assetId", assetId)
+                        .add("id", assetId)
                         .add("policy", jsonLd.compact(policy).getContent())
                 )
                 .build();
@@ -284,7 +280,7 @@ public class ConnectorRemote {
                 .add("dataDestination", destination)
                 .add("protocol", "dataspace-protocol-http")
                 .add("managedResources", false)
-                .add("assetId", assetId)
+                .add("id", assetId)
                 .add("contractId", contractAgreementId)
                 .add("connectorAddress", providerProtocolApi.toString())
                 .add("privateProperties", Json.createObjectBuilder().build())
